pyproject.toml
[project]
name = "pictopost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.8",
    "pillow>=11.2.1",
    "pillow-heif>=0.22.0",
    "streamlit>=1.45.1",
]
